Added Search Pattern, Added Transition from Wander to Search Pattern, Made Seek follow Navmesh(STILL HARDCODED)

+++ projects/App_Steering/CombinedBehaviours/FlockingSteeringBehaviors.cpp
@@ -36,7 +36,7 @@ SteeringOutput SeperationFlocking::CalculateSteering(float deltaT, SteeringAgent
 
 				m_Target.Position = neighbourPosition;
 				steering.LinearVelocity += Flee::CalculateSteering(deltaT, pAgent).LinearVelocity.GetNormalized() *
-					pow(distanceToNeigbour, -1) * pAgent->GetMaxLinearSpeed();
+					float(pow(distanceToNeigbour, -1)) * pAgent->GetMaxLinearSpeed();
 
 				/*if (pAgent->CanRenderBehavior())
 				{
